apps/paddlefold/requirements.txt
absl-py==0.13.0
biopython==1.79
dm-tree==0.1.6
docker==5.0.0
immutabledict==2.0.0
ml-collections==0.1.0
numpy==1.19.5
scipy==1.7.0
tensorboardX==2.5
